+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project *tries* to adhere to [Semantic Versioning](http://semver.org/), eve
 ## [UNRELEASED]
 - Change Item#replace method to return self
 
+## [UNRELEASED]
+- Add custom Dynomite::Errors::ValidationError and Dynomite::Errors::ReservedWordError
+  fixing rspec warnings.
+
 ## [1.2.0]
 - #7 from patchkit-net/feature/validations
 - Add a way to quickly define getters and setters using `column` method